Update binarysearch.c
You have used mid = (low+high)/2 to find the middle index calculation. This works fine. But if values of high and low are very big then the addition of high and low may overflow the highest value int can hold. 

So we use mid = low+((high-low)/2) to avoid such errors.